Kludge for Cron type events
Seems I noticed that the event emitter is not passed into rules so I
can't get the domain directly.  So in mapper I added it here so I can
get it in the match so I can 'SEE' just one time message for an event.
A kludeg for sure but it gets me around a big probmel