fix for capacity
Just a test checking.  Capacity is now passed in via the rule.  I also
had a bunch of local debugging in here I will take out when done

not sure if I need the lines at the bottom where I add the capacity from
the rule to the id-key
will take those out  if need be